chore(vouchers): integrate editmodal correct style
This commit update the `editTransaction` modal to properly integrate the
`voucher-tools` `correct` component. This tool differs from reverse tool
as it is not self contained, it requires the containing controller to
provide edited information thorugh the `source` binding. This commit
updates the edit modal to allow the user to make 'corrections' to the
transaction before submitting the correct tool with the latest changes.